# backend/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Zdravdom API
  description: Home healthcare platform API
  version: 1.0.0
  contact:
    name: Zdravdom Development Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
    UrgencyType:
      type: string
      enum: [MEDICAL_EMERGENCY, SUSPECTED_ABUSE, MEDICATION_ERROR, PATIENT_DECLINING, OTHER]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request data
        field:
          type: string
          example: email
        timestamp:
          type: string
          format: date-time
...
    GPSLocation:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
    Vitals:
      type: object
      properties:
        bloodPressure:
          type: string
          example: "120/80"
        heartRate:
          type: integer
          example: 72
        temperature:
          type: number
          format: float
          example: 36.6
        o2Saturation:
          type: integer
          example: 98

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    CompleteVisitRequest:
      type: object
      required:
        - vitals
        - clinicalNotes
      properties:
        vitals:
          $ref: '#/components/schemas/Vitals'
        clinicalNotes:
          type: string
        proceduresPerformed:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
            format: uri
        recommendations:
          type: string
        patientSignature:
          type: string
          description: Base64-encoded PNG signature
    Visit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        vitals:
          $ref: '#/components/schemas/Vitals'
        clinicalNotes:
          type: string
        proceduresPerformed:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
            format: uri
        recommendations:
          type: string
        patientSignature:
          type: string
        status:
          $ref: '#/components/schemas/VisitStatus'
        reportUrl:
          type: string
          format: uri
    EscalationRequest:
      type: object
      required:
        - urgencyType
        - notes
      properties:
        urgencyType:
          $ref: '#/components/schemas/UrgencyType'
        notes:
          type: string
        gpsLocation:
          $ref: '#/components/schemas/GPSLocation'
    Escalation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        visitId:
          type: string
          format: uuid
        urgencyType:
          $ref: '#/components/schemas/UrgencyType'
        notes:
          type: string
        gpsLocation:
          $ref: '#/components/schemas/GPSLocation'
        timestamp:
          type: string
          format: date-time
        notifiedUsers:
          type: array
          items:
            type: string

    # ─── Payments ───────────────────────────────────────────────────────────────
...
    CreateRatingRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string

# ─── Security ─────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /visits/{id}:
    get:
      summary: Get visit details
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Visit detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/start:
    put:
      summary: Provider starts visit (GPS location captured)
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Visit started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Cannot start in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/complete:
    put:
      summary: Submit visit form with vitals, notes, and signature
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteVisitRequest'
      responses:
        '200':
          description: Visit completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/escalate:
    post:
      summary: Trigger an escalation for a visit
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscalationRequest'
      responses:
        '201':
          description: Escalation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escalation'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/pdf:
    get:
      summary: Download visit report PDF
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF binary stream
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit or report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/send-to-patient:
    post:
      summary: Send visit report to patient via email
      tags: [Visits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Email queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Report email queued
        '400':
          description: Visit not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /visits/{id}/rating:
    post:
      summary: Rate a completed visit
      tags: [Ratings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRatingRequest'
      responses:
        '201':
          description: Rating submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  rating:
                    type: integer
                  review:
                    type: string
        '400':
          description: Visit not completed or already rated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Payments ─────────────────────────────────────────────────────────────────
...